Refactored some testing stuff to base classes. Added timeouts for numerous await calls in tests.
