+++ all/src/test/java/io/atomix/AtomixLeaderElectionTest.java
@@ -15,11 +15,14 @@
  */
 package io.atomix;
 
-import io.atomix.coordination.DistributedLeaderElection;
-import org.testng.annotations.Test;
-
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Function;
+
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
+import io.atomix.atomix.testing.AbstractAtomixTest;
+import io.atomix.coordination.DistributedLeaderElection;
 
 /**
  * Atomix leader election test.
@@ -28,7 +31,11 @@ import java.util.function.Function;
  */
 @Test
 public class AtomixLeaderElectionTest extends AbstractAtomixTest {
-
+  @BeforeClass
+  protected void setupCluster() throws Throwable {
+     createReplicas(5);
+  }
+  
   public void testClientLeaderElectionGet() throws Throwable {
     Atomix client1 = createClient();
     Atomix client2 = createClient();
@@ -63,7 +70,7 @@ public class AtomixLeaderElectionTest extends AbstractAtomixTest {
       resume();
     }).join();
 
-    await();
+    await(10000);
 
     election2.onElection(epoch -> {
       threadAssertTrue(epoch > lastEpoch.get());
@@ -73,7 +80,7 @@ public class AtomixLeaderElectionTest extends AbstractAtomixTest {
 
     client1.close();
 
-    await();
+    await(10000);
   }
 
 }
