+++ all/src/test/java/io/atomix/AtomixLockTest.java
@@ -15,10 +15,13 @@
  */
 package io.atomix;
 
-import io.atomix.coordination.DistributedLock;
+import java.util.function.Function;
+
+import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import java.util.function.Function;
+import io.atomix.atomix.testing.AbstractAtomixTest;
+import io.atomix.coordination.DistributedLock;
 
 /**
  * Atomix lock test.
@@ -27,7 +30,11 @@ import java.util.function.Function;
  */
 @Test
 public class AtomixLockTest extends AbstractAtomixTest {
-
+  @BeforeClass
+  protected void setupCluster() throws Throwable {
+     createReplicas(5);
+  }
+  
   public void testClientLockGet() throws Throwable {
     Atomix client1 = createClient();
     Atomix client2 = createClient();
