+++ all/src/test/java/io/atomix/AtomixMessageBusTest.java
@@ -15,11 +15,14 @@
  */
 package io.atomix;
 
-import io.atomix.catalyst.transport.Address;
-import io.atomix.coordination.DistributedMessageBus;
+import java.util.function.Function;
+
+import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import java.util.function.Function;
+import io.atomix.atomix.testing.AbstractAtomixTest;
+import io.atomix.catalyst.transport.Address;
+import io.atomix.coordination.DistributedMessageBus;
 
 /**
  * Atomix message bus test.
@@ -28,7 +31,11 @@ import java.util.function.Function;
  */
 @Test
 public class AtomixMessageBusTest extends AbstractAtomixTest {
-
+  @BeforeClass
+  protected void setupCluster() throws Throwable {
+     createReplicas(5);
+  }
+  
   public void testClientMessageBusGet() throws Throwable {
     Atomix client1 = createClient();
     Atomix client2 = createClient();
