+++ all/src/test/java/io/atomix/AtomixServerTest.java
@@ -15,9 +15,13 @@
  */
 package io.atomix;
 
+import java.util.List;
+
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import java.util.List;
+import io.atomix.atomix.testing.AbstractAtomixTest;
 
 /**
  * Atomix server test.
@@ -25,14 +29,23 @@ import java.util.List;
  * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
  */
 @Test
-public class AtomixServerTest extends AbstractServerTest {
+public class AtomixServerTest extends AbstractAtomixTest {
+  @BeforeMethod
+  protected void beforeMethod() {
+    init();
+  }
+
+  @AfterMethod
+  protected void afterMethod() {
+    cleanup();
+  }
 
   /**
    * Tests joining a server to an existing cluster.
    */
   public void testServerJoin() throws Throwable {
     createServers(3);
-    AtomixServer joiner = createServer(members, nextMember());
+    AtomixServer joiner = createServer(nextAddress(), members);
     joiner.open().thenRun(this::resume);
     await(30000);
   }
