+++ collections/src/test/java/io/atomix/collections/DistributedSetTest.java
@@ -20,7 +20,7 @@ import static org.testng.Assert.assertTrue;
 
 import org.testng.annotations.Test;
 
-import io.atomix.atomix.testing.AbstractAtomixTest;
+import io.atomix.atomix.testing.AbstractCopycatTest;
 import io.atomix.collections.state.SetState;
 import io.atomix.resource.ResourceStateMachine;
 
@@ -30,7 +30,7 @@ import io.atomix.resource.ResourceStateMachine;
  * @author <a href="http://github.com/kuujo">Jordan Halterman</a>
  */
 @Test
-public class DistributedSetTest extends AbstractAtomixTest {
+public class DistributedSetTest extends AbstractCopycatTest {
 
   @Override
   protected ResourceStateMachine createStateMachine() {
