+++ coordination/src/test/java/io/atomix/coordination/DistributedLeaderElectionTest.java
@@ -19,7 +19,7 @@ import java.util.concurrent.atomic.AtomicLong;
 
 import org.testng.annotations.Test;
 
-import io.atomix.atomix.testing.AbstractAtomixTest;
+import io.atomix.atomix.testing.AbstractCopycatTest;
 import io.atomix.coordination.state.LeaderElectionState;
 import io.atomix.copycat.client.CopycatClient;
 import io.atomix.resource.ResourceStateMachine;
@@ -30,7 +30,7 @@ import io.atomix.resource.ResourceStateMachine;
  * @author <a href="http://github.com/kuujo">Jordan Halterman</a>
  */
 @Test
-public class DistributedLeaderElectionTest extends AbstractAtomixTest {
+public class DistributedLeaderElectionTest extends AbstractCopycatTest {
 
   @Override
   protected ResourceStateMachine createStateMachine() {
@@ -68,7 +68,7 @@ public class DistributedLeaderElectionTest extends AbstractAtomixTest {
       resume();
     }).join();
 
-    await();
+    await(10000);
 
     election2.onElection(epoch -> {
       threadAssertTrue(epoch > lastEpoch.get());
@@ -78,7 +78,7 @@ public class DistributedLeaderElectionTest extends AbstractAtomixTest {
 
     client1.close();
 
-    await();
+    await(10000);
   }
 
 }
