+++ testing/pom.xml
@@ -24,6 +24,11 @@
   <dependencies>
     <dependency>
       <groupId>io.atomix</groupId>
+      <artifactId>atomix</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.atomix</groupId>
       <artifactId>atomix-resource</artifactId>
       <version>${project.version}</version>
     </dependency>
