+++ testing/src/main/java/io/atomix/atomix/testing/AbstractAtomixTest.java
@@ -15,24 +15,30 @@
  */
 package io.atomix.atomix.testing;
 
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
+import java.util.function.Function;
 
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
 
+import io.atomix.Atomix;
+import io.atomix.AtomixClient;
+import io.atomix.AtomixReplica;
+import io.atomix.AtomixServer;
 import io.atomix.catalyst.transport.Address;
 import io.atomix.catalyst.transport.LocalServerRegistry;
 import io.atomix.catalyst.transport.LocalTransport;
-import io.atomix.copycat.client.CopycatClient;
-import io.atomix.copycat.server.CopycatServer;
 import io.atomix.copycat.server.storage.Storage;
 import io.atomix.copycat.server.storage.StorageLevel;
-import io.atomix.resource.ResourceStateMachine;
+import io.atomix.resource.Resource;
+import io.atomix.resource.ResourceType;
 import net.jodah.concurrentunit.ConcurrentTestCase;
 
 /**
- * Abstract atomix test.
+ * Abstract copycat test.
  *
  * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
  */
@@ -40,108 +46,183 @@ public abstract class AbstractAtomixTest extends ConcurrentTestCase {
   protected LocalServerRegistry registry;
   protected int port;
   protected List<Address> members;
-  protected List<CopycatClient> clients;
-  protected List<CopycatServer> servers;
+  protected List<Atomix> clients;
+  protected List<Atomix> replicas;
+  protected List<AtomixServer> servers;
 
-  /**
-   * Creates a new resource state machine.
-   *
-   * @return A new resource state machine.
-   */
-  protected abstract ResourceStateMachine createStateMachine();
+  @BeforeClass
+  protected void beforeClass() {
+    init();
+  }
 
-  @BeforeMethod
+  @AfterClass
+  protected void afterClass() {
+    cleanup();
+  }
+
   protected void init() {
     port = 5000;
     registry = new LocalServerRegistry();
     members = new ArrayList<>();
     clients = new ArrayList<>();
+    replicas = new ArrayList<>();
     servers = new ArrayList<>();
   }
 
-  @AfterMethod
   protected void cleanup() {
-    clients.stream().forEach(c -> c.close().join());
-    servers.stream().forEach(s -> s.close().join());
+    clients.stream().forEach(a -> {
+      try {
+        a.close().join();
+      } catch (Exception ignore) {
+      }
+    });
+    replicas.stream().forEach(a -> {
+      try {
+        a.close().join();
+      } catch (Exception ignore) {
+      }
+    });
+    servers.stream().forEach(s -> {
+      try {
+        s.close().join();
+      } catch (Exception ignore) {
+      }
+    });
+
+    clients.clear();
+    replicas.clear();
+    servers.clear();
   }
-  
+
+  /**
+   * Creates a resource factory for the given type.
+   */
+  @SuppressWarnings("unchecked")
+  protected <T extends Resource<T>> Function<Atomix, T> get(String key, ResourceType<?> type) {
+    return a -> {
+      try {
+        return a.get(key, (ResourceType<T>) type).get();
+      } catch (InterruptedException | ExecutionException e) {
+        throw new RuntimeException(e);
+      }
+    };
+  }
+
+  /**
+   * Creates a resource factory for the given type.
+   */
+  @SuppressWarnings("unchecked")
+  protected <T extends Resource<T>> Function<Atomix, T> create(String key, ResourceType<?> type) {
+    return a -> {
+      try {
+        return a.create(key, (ResourceType<T>) type).get();
+      } catch (InterruptedException | ExecutionException e) {
+        throw new RuntimeException(e);
+      }
+    };
+  }
+
   /**
    * Returns the next server address.
    *
    * @return The next server address.
    */
-  private Address nextAddress() {
-    Address address = new Address("localhost", port++);
-    members.add(address);
-    return address;
+  protected Address nextAddress() {
+    return new Address("localhost", port++);
   }
 
   /**
-   * Creates a set of Raft servers.
+   * Creates a client.
    */
-  protected List<CopycatServer> createServers(int nodes) throws Throwable {
-    List<CopycatServer> servers = new ArrayList<>();
-
-    List<Address> members = new ArrayList<>();
-    for (int i = 0; i < nodes; i++) {
-      members.add(nextAddress());
-    }
-
-    for (int i = 0; i < nodes; i++) {
-      CopycatServer server = createServer(members.get(i));
-      server.open().thenRun(this::resume);
-      servers.add(server);
-    }
-
-    await(0, nodes);
-
-    return servers;
+  protected Atomix createClient() throws Throwable {
+    Atomix client = AtomixClient.builder(members).withTransport(new LocalTransport(registry)).build();
+    client.open().thenRun(this::resume);
+    clients.add(client);
+    await(10000);
+    return client;
   }
 
   /**
-   * Creates a set of Raft servers.
+   * Creates an Atomix replica.
    */
-  protected List<CopycatServer> createServers(int live, int total) throws Throwable {
-    List<CopycatServer> servers = new ArrayList<>();
-
-    List<Address> members = new ArrayList<>();
-    for (int i = 0; i < total; i++) {
-      members.add(nextAddress());
-    }
-
-    for (int i = 0; i < live; i++) {
-      CopycatServer server = createServer(members.get(i));
-      server.open().thenRun(this::resume);
-      servers.add(server);
-    }
-
-    await(0, live);
-
-    return servers;
+  protected Atomix createReplica(Address address, List<Address> members) {
+    AtomixReplica replica = AtomixReplica.builder(address, members)
+        .withTransport(new LocalTransport(registry))
+        .withStorage(Storage.builder()
+            .withStorageLevel(StorageLevel.MEMORY)
+            .withMaxEntriesPerSegment(8)
+            .withMinorCompactionInterval(Duration.ofSeconds(5))
+            .withMajorCompactionInterval(Duration.ofSeconds(10))
+            .build())
+        .build();
+    replicas.add(replica);
+    return replica;
   }
 
   /**
-   * Creates a Raft server.
+   * Creates an Atomix server.
    */
-  protected CopycatServer createServer(Address address) {
-    CopycatServer server = CopycatServer.builder(address, members)
-      .withTransport(new LocalTransport(registry))
-      .withStorage(new Storage(StorageLevel.MEMORY))
-      .withStateMachine(createStateMachine())
-      .build();
+  protected AtomixServer createServer(Address address, List<Address> members) {
+    AtomixServer server = AtomixServer.builder(address, members)
+        .withTransport(new LocalTransport(registry))
+        .withStorage(Storage.builder()
+            .withStorageLevel(StorageLevel.MEMORY)
+            .withMaxSegmentSize(1024 * 1024)
+            .withMaxEntriesPerSegment(8)
+            .withMinorCompactionInterval(Duration.ofSeconds(3))
+            .withMajorCompactionInterval(Duration.ofSeconds(7))
+            .build())
+        .build();
     servers.add(server);
     return server;
   }
 
   /**
-   * Creates a Copycat client.
+   * Creates a set of Atomix instances.
    */
-  protected CopycatClient createClient() throws Throwable {
-    CopycatClient client = CopycatClient.builder(members).withTransport(new LocalTransport(registry)).build();
-    client.open().thenRun(this::resume);
-    await();
-    clients.add(client);
-    return client;
+  protected List<Atomix> createReplicas(int nodes) throws Throwable {
+    List<Address> members = new ArrayList<>();
+    for (int i = 0; i < nodes; i++) {
+      members.add(nextAddress());
+    }
+    this.members.addAll(members);
+
+    List<Atomix> replicas = new ArrayList<>();
+    for (int i = 0; i < nodes; i++) {
+      Atomix atomix = createReplica(members.get(i), members);
+      atomix.open().thenRun(this::resume);
+      replicas.add(atomix);
+    }
+
+    await(0, nodes);
+    return replicas;
   }
 
+  /**
+   * Creates a set of Raft servers.
+   */
+  protected List<AtomixServer> createServers(int live, int total) throws Throwable {
+    List<Address> members = new ArrayList<>();
+    for (int i = 0; i < total; i++) {
+      members.add(nextAddress());
+    }
+    this.members.addAll(members);
+
+    List<AtomixServer> servers = new ArrayList<>();
+    for (int i = 0; i < live; i++) {
+      AtomixServer server = createServer(members.get(i), members);
+      server.open().thenRun(this::resume);
+      servers.add(server);
+    }
+
+    await(0, live);
+    return servers;
+  }
+
+  /**
+   * Creates a set of Raft servers.
+   */
+  protected List<AtomixServer> createServers(int nodes) throws Throwable {
+    return createServers(nodes, nodes);
+  }
 }
